configured shadcn to work in a shared library -- finalllly!!
