+++ clients/core/src/index.js
@@ -1,4 +1,4 @@
-import './style.css'
+import '@/src/index.css'
 
 // Define the URLs here for the webpack config to find the remote modules
 window.templateComponent2Url = 'http://localhost:3001'
