+++ clients/shared_library/components.json
@@ -11,11 +11,12 @@
       "prefix": ""
     },
     "aliases": {
-      "components": "@/components",
-      "utils": "@/lib/utils",
-      "ui": "@/components/ui",
-      "lib": "@/lib",
-      "hooks": "@/hooks"
+      "components": "components",
+      "utils": "lib/utils",
+      "ui": "components/ui",
+      "lib": "lib",
+      "hooks": "hooks"
     },
     "iconLibrary": "lucide"
-  }
+}
+
