+++ clients/shared_library/eslint.config.mjs
@@ -1,0 +1,10 @@
+import rootConfig from '../eslint.config.mjs'
+
+export default [
+  ...rootConfig,
+  {
+    // Optionally add any subfolder-specific rules or settings
+    files: ['**/*.ts', '**/*.tsx'],
+    rules: {},
+  },
+]
