--- clients/shared_library/index.js
@@ -1,1 +1,0 @@
-// all components, that shall be exported need to mentioned here
