+++ clients/shared_library/package.json
@@ -4,7 +4,21 @@
     "private": true,
     "main": "./index.js",
     "dependencies": {
-        "react": "^18.3.1"
+        "@radix-ui/react-slot": "^1.1.0",
+        "@shadcn/ui": "^0.0.4",
+        "class-variance-authority": "^0.7.0",
+        "clsx": "^2.1.1",
+        "lucide-react": "^0.460.0",
+        "react": "^18.3.1",
+        "tailwind-merge": "^2.5.4",
+        "tailwindcss-animate": "^1.0.7"
     },
-    "packageManager": "yarn@4.1.0"
+    "exports": {
+        "./components/*": "./components/*",
+        "./src/*": "./src/*"
+    },
+    "packageManager": "yarn@4.1.0",
+    "devDependencies": {
+        "tailwindcss": "^3.4.15"
+    }
 }
